+++ src/CodecTactics.Core/Network/NodeState.cs
@@ -2,12 +2,15 @@ namespace CodecTactics.Core.Network;
 
 public sealed class NodeState
 {
-    public NodeState(NodeId id)
+    public NodeState(NodeId id, NodeType type = NodeType.Standard)
     {
         Id = id;
+        Type = type;
     }
 
     public NodeId Id { get; }
+
+    public NodeType Type { get; }
 
     public NodeOwner Owner { get; private set; } = NodeOwner.Neutral;