+++ src/CodecTactics.Core/Network/NodeType.cs
@@ -1,0 +1,9 @@
+namespace CodecTactics.Core.Network;
+
+public enum NodeType
+{
+    Standard,
+    Resource,
+    Relay,
+    Firewall
+}